Cast to unsigned in HpackVarintRoundTripTest.
Chromium's string uses unsigned char, Google3 uses signed.  This causes an
assertion in HpackVarintRoundTripTest to fail.  This was not an issue when
prefixes were only allowed to be up to 7 bits long, because then prefix_mask was
at most 127.  8 bit long prefixes were allowed at cr/225873577, which was then
merged to QUICHE at
https://quiche.googlesource.com/quiche/+/cc66370c8f56ee74aa9f13407249a836ea26b6ce
When trying to roll this into Chromium, HpackVarintRoundTripTests fail, see
trybot output at https://crrev.com/c/1391715.

I locally reproduced the failure in my Chromium checkout and verified that this
change fixes it.

PiperOrigin-RevId: 227084858
Change-Id: Ifcea92d2234d87dabba6ce82b2dac1e3a5fd90a8