Automated rollback of changelist 227019741.  See https://crbug.com/921074
*** Reason for rollback ***

It appears that the caller does not expect the method to return an error, which results in weird behavior on 32-bit platforms

*** Original change description ***

Avoid truncation error on 32 bit platforms.

This is so that truncation warning can be turned into compiler errors for Blink
on 32-bit Android platforms.  See https://crbug.com/916990 for more context.

gfe-relnote: No functional change on 64-bit platforms.  Not flag protected.

***

PiperOrigin-RevId: 229219479
Change-Id: Ie3ef2d518951894f4a1ee8a7b92478f3a946dfa9