De-pointerize QuicServerConfigProtobuf variables
This CL converts a bunch of unique_ptr<QuicServerConfigProtobuf> objects to plain QuicServerConfigProtobuf.  The extra indirection is not helping us here, and just introduces a bunch of "but what if it's null?" confusion.  Especially since protobufs are moveable now, there's almost no performance implication to this change, and none of this code is in the serving path anyhow.

gfe-relnote: Use plain objects instead of unique_ptrs.  No functional change intended, not flag-protected.
PiperOrigin-RevId: 239653446
Change-Id: I9007e89b33dfe201354d89fdbaa4d12c5d86fbed