Add a new implementation of SpdyUnsafeArena for code sharing between chromium and envoy. This new implementation is merely a copy of SpdyUnsafeArenaImpl from chromium (https://cs.chromium.org/chromium/src/net/spdy/platform/impl/spdy_unsafe_arena_impl.h?g=0&l=17) which itself is an older data structure used by SpdyHeaderBlock.
This new implemenation is not used as google3 impl. However, it's nice to have
it in quiche for code sharing between platforms.

Once this class is in quiche, chromium platform can point to this one instead of keeping the same code in SpdyUnsafeArenaImpl.

gfe-relnote: n/a (new class not in use)
PiperOrigin-RevId: 243133213
Change-Id: I52875503efc6be97c7a531d9c3e02780d66b88cc