Add functions to enable/disable QuartcPeer and obtain the last sequence numbers sent by each of its sources.
These functions allow a test to stop sending and drain network traffic before
closing the connection, by disabling both peers and waiting until each receives
the sequence number last sent by the other.

Note that this might not work perfectly on lossy networks, as the last message
may be dropped.  In this case, draining should use a timeout.  For example, the
receiver may stop waiting based if it does not receive any frames for an RTT.

gfe-relnote: n/a (Quartc test-only)
PiperOrigin-RevId: 243880014
Change-Id: I17eb39fe9db65138094be885ba2330936dbeac4e