Prevent include failures of quic_flag_utils
We've had quiche merge failures due to missing includes because in google3 quic_flags ends up including quic_flag_utils but not in Chromium. This ensures inclusion in the api header to make sure behavior is consistent between google3 and Chromium. This is also useful because it is very common for code that uses flags to also want to use flag counts.

gfe-relnote: n/a, compile-only
PiperOrigin-RevId: 248687080
Change-Id: I0a163bb06d4b4a8ecc1747245214f329ba565ba3