Prepare tests and fuzzer for client connection IDs
When working on client connection IDs, we observed that the fuzzer test was incorrectly deciding when to call TestConnectionIdToUInt64(). Long term, we will make the fuzzer support variable length connection IDs and remove all calls to TestConnectionIdToUInt64 there but for now the fuzzer should only call that when the connection ID is not empty.

We also observed that there were old end to end tests that tested Google-QUIC Connection ID compression and were erroneously not disabled in v44. We've tweaked these tests over the past months to accommodate new features but at this point the correct path forward is to instead make them specific to what they're testing which is v<=43.

gfe-relnote: n/a, test-only
PiperOrigin-RevId: 250388286
Change-Id: I03181159817a458f9ba7893b8b592b28be4d8865