Add a client connection ID to version negotiation packet creation
This CL adds a new client_connection_id parameter to the functions that build version negotiation packets, and pass it through to AppendIetfConnectionIds instead of EmptyQuicConnectionId. All callers of this function call it with EmptyQuicConnectionId. To verify this we have a DCHECK that makes sure client_connection_id is empty when the disabled flag quic_do_not_override_connection_id is false. This CL also adds QuicVersion::SupportsClientConnectionIds which subsequent CLs will build on.

gfe-relnote: no functional change, add function parameter that is always empty, not flag-protected
PiperOrigin-RevId: 250551156
Change-Id: I378c67df16fc262145c5c88a76ca5440f72a29d8