Make dispatcher VN client CID test work for all flag values
We noticed that QuicDispatcherTest.StatelessVersionNegotiationWithClientConnectionId fails on Chromium because quic_allow_variable_length_connection_id_for_negotiation is false. This uncovered that this test relies on that flag being true. To avoid issues in the future if we have to flip this flag back, this CL changes this test to set the flag to true.

gfe-relnote: n/a, test-only
PiperOrigin-RevId: 252026292
Change-Id: I63c928e3b174cfff0c71e9fc4a971ff6d2a7ffd0