Make QuicSpdyStream::OnStreamHeaderList() return void.
Only QuicSpdyStream::OnHeadersFrameEnd() cares about the return value.
When blocked encoding is implemented, the return value will be
considered and passed back to HttpDecoder in the synchronous case, but
will be ignored in the async case, where OnStreamHeaderList() will be
called not from inside an HttpDecoder::ProcessInput() call.

gfe-relnote: n/a, change to QUIC v99-only code path.
PiperOrigin-RevId: 255242560
Change-Id: I4ef67a7f8b0926830e12a3da0f6317b32db951e3