In HTTP/3, write Priority on control stream before writing headers.
Currently the priority is in gQUIC fashion, where we only use the weight to build priority queue.

Priority might be sent on the control stream before handshake is confirmed. So this CL modifies QuicSendControlStream to send stream type and settings before any other data is sent.

gfe-relnote: v99 only, not in prod.
PiperOrigin-RevId: 256285943
Change-Id: Iaf72f9d6256a2692b284e3f6142755ff0d04d710