Remove stream counting for pending streams.
The direct reason for this CL is that num_dynamic_incoming_streams_ is only decremented but not incremented for pending streams. This will result in incorrect accounting of streams.

But in general, pending streams are not known to be dynamic or static, so we shouldn't count them before activating them.

gfe-relnote: v99 only, not used in prod.
PiperOrigin-RevId: 258493376
Change-Id: Id68e0f9a69dbb6101f2cf817afe0f111cc19ba07