Put Google QUIC Connection Close code in the extracted code field
When processing Google QUIC Connection Close frames, place a copy
of the close code into the QuicConnectionCloseFrame::extracted_error_code
field. In this way, as we migrate through a hybrid Google/IETF QUIC
implementation, the Google QUIC close code is always available in
this field, regardless of the QUIC type in use.  This does not change
the existing code's placement of the error code in
QuicConnectionCloseFrame::quic_error_code.

gfe-relnote: N/A, part of a revision of the ConnectionCloseFrame in an earlier CL.
PiperOrigin-RevId: 262114442
Change-Id: Iaa472c1e040dd7de4d5b6a3c2484d5aad1408413