Change Application and Transport ConnectionClose codes to uint64_t
IETF QUIC Transport and Application ConnectionClose Error Codes were 16 bit
unsigned numbers. Draft 22 of the IETF QUIC specification changed them to be a
varint62, which we keep internally as a uint64_t. This CL effects that change.

The benefits of this CL are A) somewhat cleaner internal code (elimination of random casts) and B) applications can use the full 62 bits of the connection close codes when using IETF QUIC.

gfe-relnote: N/A IETF QUIC only, V99 flag protected.
PiperOrigin-RevId: 267594523
Change-Id: I0d6a8af0bb134c15d6b034de4237a02e0458ee47