BEGIN_PUBLIC Update uses of INSTANTIATE_TEST_SUITE_P and TEST_P that leave the arguments blank to pass a non empty value (e.g. All). END_PUBLIC
tl;dr; there are a number of places where INSTANTIATE_TEST_SUITE_P or TEST_P is used with empty arguments, which causes problems with some planned edits to googletest.

Parameterized macro test names seem to be broken. That is if you try to use a macro invocation as an argument to one of the TEST_P or INSTANTIATE_TEST_SUITE_P macros, no error is generated but it seems the resulting test won?t actually get run. This turns out to be a side effect of doing ?stringification? wrong. But when that is fixed (cl/263590733), some compilers reject the resulting code in the case where the newly stringified argument is empty (I?ve seen at least MSVC reporting errors).

This was constructed via a CL (like cl/266814595 using go/qnqcj) that fixes the naming issue and inserts checks for empty names, and then fixing the cases that finds.

go/cleanup-139533596

Tested:
    TAP --sample ran all affected tests and none failed
    http://test/OCL:270476273:BASE:270469786:1569104339530:7b684d73
PiperOrigin-RevId: 270650287
Change-Id: I6884ffd68843f522106ca01eef1a35579baab062