Fix auto-setting IETF stream flow control limits
The code would set the IETF limits in the old setter in the hopes of having them used when they weren't individually set. But since they were set in the constructor that didn't work. The new code checks the old value from the getter, which gets us the requested behavior of using the old setter value in the absence of other information.

gfe-relnote: fix IETF flow control, protected by disabled TLS flag
PiperOrigin-RevId: 274898114
Change-Id: I77ddd45b00715a15cc301c4f48833a28184b8685