Do not track unacknowledged references on the encoder stream in QpackBlockingManager.
This CL essentially reverts change 263116461, which was an oversight on my part.
The reason is that unacknowledged references only have to be tracked so that no
entry with unacknowledged references is evicted from the dynamic table.
However, eviction is caused by an insertion on the encoder stream, and by that
time that is processed, it is guaranteed that every reference on the encoder
stream sent before that insertion is already processed by the decoder.  (And of
course there can not be references on the encoder stream after the entry is
evicted.)  So references on the encoder stream do not matter, only references on
other streams.

gfe-relnote: n/a, change to QUIC v99-only code.  Protected by existing disabled gfe2_reloadable_flag_quic_enable_version_99.
PiperOrigin-RevId: 275309366
Change-Id: Id3e4845f76815bc228b567fb55124a5d771843d9