Add QUIC_EXPORT_PRIVATE to prepare for presubmit
A subsequent CL (cl/276207410) will add a presubmit to mitigate QUICHE merge issues, this CL gets our code-base to a state of compliance with that presubmit. These haven't caused issues in Chromium if they were only used in the same build unit, but it's worth it to have a consistent use across the codebase to make sure everything works in Chromium.

gfe-relnote: no behavior change, QUIC_EXPORT_PRIVATE is a no-op in google3
PiperOrigin-RevId: 276276729
Change-Id: I67ab8c9df92d24d0006de7084ca563153f0bf07f