Fix some test expectation regarding SendControlFrame().
Before IETF QUIC, only 1 control frame will be sent for a stream reset. But IETF requires both a RST and STOP_SENDING. The tests that explicitly expect SendControlFrame() to be called once is making the wrong assumption.

The reason they still pass is the first write was buffered so the second write won't trigger.

Remove the explicit Times(1) to avoid confusion.

gfe-relnote: n/a. test only.
PiperOrigin-RevId: 276736424
Change-Id: Iffa8de79c51e28355ef4aa8381ae049948e40b67