Simplify ConstructMisFramedEncryptedPacket() signature.
Both call sites have only one version, so taking a ParsedQuicVersion is a better
choice than taking a ParsedQuicVersionVector.

This is motivated by the variable |version| being unused in
ConstructMisFramedEncryptedPacket(), which causes compilation failure in Envoy,
but once I'm at it, it makes sense to make this small improvement.

gfe-relnote: n/a, test-only change.
PiperOrigin-RevId: 277916457
Change-Id: Ica41b8985510bddbe81dadf9843e0a9620ef3ffe