gfe-relnote: (n/a) In QuicTimeAccumulator, change a "static constexpr data member" to a "static constexpr function". No behavior change, not protected.
This caused a merge error in Chromium because QuicTimeAccumulator::kNotRunningSentinel does not have a definition. The requirement for a definition has been removed since c++17.

I could not think of a clean way to prevent this in presubmit, simply ban "static constexpr QuicTime" seems too restrictive, because some code may want to use it with a definition in .cc.

PiperOrigin-RevId: 282603664
Change-Id: Ia677ca93b851725e4509c4960c5e71fdeac5f903