Switch the order of QPACK streams sending stream type.
QPACK decoder stream is created before the encoder stream. It would be easier if they send out stream types in the order they are created, so that when the session is write blocked and resumed, the same order of sending can be expected.

Also removed outdated TODO's. A bigger and better change can be made when we support 0 initial stream limit.

gfe-relnote: protected by disabled v99 flag.
PiperOrigin-RevId: 282818824
Change-Id: I4f998fd186669dabaa806173c078300396e0a5e2