Set quic_send_max_push_id_with_settings in some test cases
These test cases don't work with the combination of flags where this flag
is off and the v99 flag is on. (They do work if both are off or both are
on.) The default in chromium is that v99 is on but this flag is off; always
having this flag on fixes the tests for the chromium merge.

gfe-relnote: n/a (flip a flag in some tests)
PiperOrigin-RevId: 283869287
Change-Id: I7b284351df783e781f3070122451d858397dd208