Introduce MASQUE, part 1: shared code
This CL introduces the shared code for MASQUE as defined by <https://tools.ietf.org/html/draft-schinazi-masque>. Most of the work here is plumbing in order to override the right methods of the QUIC codebase. The meat of the MASQUE protocol work is in masque_compression_engine.cc.

gfe-relnote: n/a, adds unused code
PiperOrigin-RevId: 285439559
Change-Id: I2739a002a723d5ba783034fb78959f02f3d89365