Remove the DCHECK from ParsedQuicVersion constructor.
The DCHECK causes the compiler to decide that ParsedQuicVersion constructor is not constexpr, which in turn causes Chromium code to fail to compile in various places (partially because some of our version constants are constexpr, partially because the DCHECK gets run at initialization time and Chromium bans initalization-time constructors).  As far as I can tell, this is because ParsedQuicVersionIsValid is not constexpr, though I suspect the DCHECK itself is not constexpr either.

gfe-relnote: n/a (no functional change)
PiperOrigin-RevId: 289938397
Change-Id: Ib3d03b4be710aea65a75993d7b5d49837c3e420e