gfe-relnote: Deprecate unverified flag quic_bbr_less_probe_rtt false.
This is currently only used by Quartc, but we're hoping to migrate to BBRv2, where these connection options currently have no effect and some of the options are implemented in part(ie: Only reduce CWND to 1/2 BDP for PROBE_RTT).

If the COPTs behind this flag are used or planned to be used in production, then I can delay this CL until BBRv2 is 'done', but ideally I'd like to clean them up now if I can.

PiperOrigin-RevId: 290152900
Change-Id: If18f613f08ec1d44cd54d1857f64fda66bb90cf3