gfe-relnote: Fix invalid downcast.
QuicTransportServerSessionTest fails on UBSAN builds in Chromium, see
https://crrev.com/c/2015242.

QuicTransportServerSession initializes |crypto_stream_| in the constructor using
CreateCryptoServerStream(), which since cr/290819354 might return a
TlsServerHandshaker, which is a derived class of QuicCryptoServerStreamBase but
not of QuicCryptoServerStream.  Hence the downcast fails in
QuicTransportServerSessionTest constructor.  Fortunately
QuicTransportServerSessionTest does not need a QuicCryptoServerStream, only a
QuicCryptoServerStreamBase: it calls OnSuccessfulVersionNegotiation() which is a
QuicCryptoStream method, and passes it to  HandshakeWithFakeClient() which takes
a QuicCryptoServerStreamBase*.
PiperOrigin-RevId: 291203032
Change-Id: I7b3ec41c64eeb46074ee6aa98bff7bd132a20daa