Fix connection close check in quic_client_interop_test_bin
The current connection close check only works if the server sends a
connection close in response to ours. The transport draft doesn't require
that a peer send a connection close in response to a connection close
(though it is the only frame type that can be sent once a connection enters
the closing state). It seems perfectly reasonable to declare the connection
close procedure has completed with a zero error code if we send a
connection close with a zero error code.

gfe-relnote: n/a (test-only tool change)
PiperOrigin-RevId: 293557827
Change-Id: I1fe90fd65703f061e7d920fb7f20380c0ed38162