Change number of connection attempts in QuicClientBase::Connect
This removes its dependency on GetNumSentClientHellos. Though instead of
comparing the number of connection attempts to kMaxClientHellos, it could
probably just try twice - once for whatever initial version the client is
configured for, and a second time for a new version in response to version
negotiation.

gfe-relnote: n/a (tools only change)
PiperOrigin-RevId: 295235905
Change-Id: Ia1d601bf66f61639cf45276534be122d63f5e1d3